requirements.txt
streamlit <= 1.36.0
pandas <= 2.2.2
plotly <= 5.22.0
mysql-connector-python <= 9.0.0